переписан конструктор GearKit

+++ lab2_20/lab024/GeneratePrimes.h
@@ -3,5 +3,5 @@
 #include <set>
 #include <vector>
 
-std::vector<bool> SieveResult(unsigned int bound); // позиции простых по решету Эратосфена
-std::set<int> GeneratePrimeNumbersSet(unsigned int upperBound); // множество простых, не больших upperBound
+std::vector<bool> SieveResult(size_t bound); // позиции простых по решету Эратосфена
+std::set<size_t> GeneratePrimeNumbersSet(size_t upperBound); // множество простых, не больших upperBound
